type casting in SQL + added totalInvested on running transactions!

+++ sql/schema/002_stocks.sql
@@ -2,8 +2,8 @@
 CREATE TABLE stocks(
     symbol TEXT PRIMARY KEY,
     company_name TEXT NOT NULL,
-    current_price DECIMAL(10, 2) NOT NULL,
-    previous_close DECIMAL(10, 2),
+    current_price DOUBLE PRECISION NOT NULL,
+    previous_close DOUBLE PRECISION,
     updated_at TIMESTAMP NOT NULL
 );
